fix forwarding edge case for vim split
Steps to reproduce issue
- Split tmux in 2 vertical panes
- In the right most pane, open vim
- Create a horizontal split in vim
- Try to resize left and right
Expected: Tmux separator moves left and right
Actual: Tmux separator moves only left

The issue was when considering neighbouring vim splits.
The previous logic covered the case where we are the last split of the
last pane in either axis. In this case we should resize vim.
But this should only be done if there is a vim split preceeding along
the axis.